Add new assets and update avatar with debug unlock
Added multiple image and SVG assets for branding and icons. Updated the profile avatar in index.html to use a new image instead of an inline SVG. Enhanced the debug logger in script.js to buffer messages and unlock on 5 avatar clicks, with a new setupAvatarDebugUnlock function. Updated styles to support the new avatar image.